+++ scopesim/effects/effects_utils.py
@@ -1,7 +1,5 @@
 from copy import deepcopy
 
-import numpy as np
-
 from .. import effects as efs
 from ..optics.radiometry_utils import empty_surface_list
 
@@ -57,77 +55,3 @@ def is_spectroscope(effects):
                          for eff in effects])
 
     return bool(has_apertures and has_trace_lists)
-
-
-def atmospheric_refraction(lam, z0=60, temp=0, rel_hum=60, pres=750,
-                           lat=-24.5, h=3064):
-    """Compute atmospheric refraction
-
-    The function computes the angular difference between the apparent position
-    of a star seen from the ground and its true position.
-
-    Parameters
-    ----------
-    lam : float, np.ndarray
-        [um] wavelength bin centres
-    z0 : float, optional
-        [deg] zenith distance. Default is 60 deg from zenith
-    temp : float, optional
-        [deg C] ground temperature. Default is 0 deg C
-    rel_hum : float, optional
-        [%] relative humidity. Default is 60%
-    pres : float, optional
-        [millibar] air pressure. Default is 750 mbar
-    lat : float, optional
-        [deg] latitude. Default set for Cerro Armazones: 24.5 deg South
-    h : float, optional
-        [m] height above sea level. Default is 3064 m
-
-    Returns
-    -------
-    ang : float, np.ndarray
-        [arcsec] angle between real position and refracted position
-
-    References
-    ----------
-    See Stone 1996 and the review by S. Pedraz -
-    http://www.caha.es/newsletter/news03b/pedraz/newslet.html
-    """
-
-    # need T, P, RH for Ps, Pw Pa
-    T = 273.15 + temp
-
-    Ps = -10474. + 116.43 * T - 0.43284 * T**2 + 0.0005384 * T**3
-    Pw = Ps * rel_hum / 100.
-    Pa = pres - Pw
-
-    # need n0 for gamma
-    sig = 1. / lam
-    Da = (Pa / T) * (1. + Pa * (57.9E-8 - 0.0009325 / T + 0.25844 / T**2))
-    Dw = (Pw / T) * (1. + Pw * (1. + 3.7E-4 * Pw) *
-                     (-2.37321E-3 + 2.23366 / T - 710.792 / T**2 +
-                      77514.1 / T**3))
-
-    na = Da * (2371.34 + 683939.7 / (130. - sig**2) + 4547.3 / (38.9 - sig**2))
-    nw = Dw * (6487.31 + 58.058 * sig**2 - 0.7115 * sig**4 + 0.08851 * sig**6)
-    n0 = 1E-8 * (na + nw) + 1.
-
-    # need gamma, kappa and beta for R
-    g = n0 - 1.
-    b = 0.001254 * (273.15 + temp) / 273.15
-    k = 1. + 0.005302 * np.sin(np.deg2rad(lat))**2 \
-        - 0.00000583 * np.sin(2. * np.deg2rad(lat))**2 - 0.000000315 * h
-
-    R = k * g * (1 - b) * np.tan(np.deg2rad(z0)) \
-        - k * g * (b - g/2.) * np.tan(np.deg2rad(z0))**3
-
-    # the refraction is the side of a triangle, although the triangle
-    # side makes an arc across the sky.
-    # We want the angle that this triangle side is subtending
-    # Using the small angle approximation (which is in radians),
-    # we can get the angle of refraction
-
-    ang = np.rad2deg(R) * 3600
-
-    # return value is in arcsec
-    return ang
